+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.3.0
+
+### Changed
+- `Paysera/php_basic_feature_checking_explicitly` fixer now changes `isset` if `empty` checks array key, i.e.:
+`!empty($a['b']) -> isset($a['b'])`
+`!empty($a) -> count($a) > 0`
+`empty($a) -> count($a) === 0`
+
+
 ## 1.2.0
 
 ### Fixed